+++ demo/src/test/java/test/base/BaseFrame.java
@@ -2,7 +2,7 @@ package test.base;
 
 import com.formdev.flatlaf.FlatLaf;
 import com.formdev.flatlaf.fonts.roboto.FlatRobotoFont;
-import com.formdev.flatlaf.themes.FlatMacDarkLaf;
+import com.formdev.flatlaf.themes.FlatMacLightLaf;
 
 import javax.swing.*;
 import java.awt.*;
@@ -19,7 +19,7 @@ public class BaseFrame extends JFrame {
     protected static void installLaf() {
         FlatRobotoFont.install();
         FlatLaf.registerCustomDefaultsSource("raven.modal.demo.themes");
-        FlatMacDarkLaf.setup();
-        UIManager.put("defaultFont", new Font(FlatRobotoFont.FAMILY, Font.PLAIN, 13));
+        FlatMacLightLaf.setup();
+        // UIManager.put("defaultFont", new Font(FlatRobotoFont.FAMILY, Font.PLAIN, 13));
     }
 }
