+++ demo/src/test/java/test/validation/ValidationDrawerBuilder.java
@@ -4,8 +4,8 @@ import com.formdev.flatlaf.FlatClientProperties;
 import raven.extras.AvatarIcon;
 import raven.modal.demo.forms.*;
 import raven.modal.drawer.DrawerPanel;
-import raven.modal.drawer.data.Item;
-import raven.modal.drawer.data.MenuItem;
+import raven.modal.drawer.item.Item;
+import raven.modal.drawer.item.MenuItem;
 import raven.modal.drawer.menu.*;
 import raven.modal.drawer.simple.SimpleDrawerBuilder;
 import raven.modal.drawer.simple.footer.SimpleFooterData;
