+++ modal-dialog/src/main/java/raven/modal/drawer/item/Item.java
@@ -1,4 +1,4 @@
-package raven.modal.drawer.data;
+package raven.modal.drawer.item;
 
 import java.util.ArrayList;
 import java.util.List;
