+++ modal-dialog/src/main/java/raven/modal/drawer/item/MenuItem.java
@@ -1,4 +1,4 @@
-package raven.modal.drawer.data;
+package raven.modal.drawer.item;
 
 /**
  * @author Raven
